+++ executor-service/src/main/java/executor/service/model/dto/Step.java
@@ -1,4 +1,4 @@
-package executor.service.model;
+package executor.service.model.dto;
 
 import jakarta.persistence.Embeddable;
