+++ executor-service/src/main/java/executor/service/model/dto/ThreadPoolConfig.java
@@ -1,4 +1,4 @@
-package executor.service.model;
+package executor.service.model.dto;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
