+++ executor-service/src/main/java/executor/service/service/proxy/ProxySourcesClient.java
@@ -1,6 +1,6 @@
 package executor.service.service.proxy;
 
-import executor.service.model.ProxyConfigHolder;
+import executor.service.model.dto.ProxyConfigHolder;
 
 public interface ProxySourcesClient {
     ProxyConfigHolder getProxy();
