+++ executor-service/src/main/java/executor/service/service/scenario/ScenarioService.java
@@ -1,6 +1,6 @@
 package executor.service.service.scenario;
 
-import executor.service.model.Scenario;
+import executor.service.model.dto.Scenario;
 
 public interface ScenarioService {
