+++ executor-service/src/main/java/executor/service/service/step/impl/StepExecutionClickXpath.java
@@ -1,6 +1,6 @@
 package executor.service.service.step.impl;
 
-import executor.service.model.Step;
+import executor.service.model.dto.Step;
 import executor.service.service.step.AbstractStepExecution;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
