+++ executor-service/src/main/java/executor/service/service/webDriver/WebDriverInitializer.java
@@ -1,6 +1,6 @@
 package executor.service.service.webDriver;
 
-import executor.service.model.ProxyConfigHolder;
+import executor.service.model.dto.ProxyConfigHolder;
 import org.openqa.selenium.WebDriver;
 
 public interface WebDriverInitializer {
