+++ executor-service/src/test/java/executor/service/model/StepTest.java
@@ -1,5 +1,6 @@
 package executor.service.model;
 
+import executor.service.model.dto.Step;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
