+++ executor-service/src/test/java/executor/service/model/WebDriverConfigTest.java
@@ -1,5 +1,6 @@
 package executor.service.model;
 
+import executor.service.model.dto.WebDriverConfig;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
