+++ executor-service/src/test/java/executor/service/params/ActionsArgumentsProvider.java
@@ -1,6 +1,6 @@
 package executor.service.params;
 
-import executor.service.model.Step;
+import executor.service.model.dto.Step;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.ArgumentsProvider;
