+++ executor-service/src/test/java/executor/service/params/ScenariosArgumentsProvider.java
@@ -1,6 +1,6 @@
 package executor.service.params;
 
-import executor.service.model.Scenario;
+import executor.service.model.dto.Scenario;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.ArgumentsProvider;
