+++ executor-service/src/test/java/executor/service/service/proxy/ProxySourcesClientServiceTest.java
@@ -1,6 +1,6 @@
 package executor.service.service.proxy;
 
-import executor.service.model.ProxyConfigHolder;
+import executor.service.model.dto.ProxyConfigHolder;
 import executor.service.util.file.FileParser;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
